requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langchain-community
langchain-huggingface
chromadb
pydantic
python-multipart
sentence-transformers
